+++ src/KairosWeb_Groep6/Models/KairosViewModels/Baten/MedewerkerNiveauBaatViewModel.cs
@@ -17,19 +17,17 @@ namespace KairosWeb_Groep6.Models.KairosViewModels.Baten
 
         [HiddenInput]
         public Soort Soort { get; set; }
-
-        //public string Beschrijving { get; set; }
-
+        
         [Required(ErrorMessage = "Gelieve een aantal uren op te geven")]
         [Range(0, double.MaxValue, ErrorMessage = "Gelieve een positief getal voor het aantal uur op te geven.")]
-        public double Uren { get; set; }
+        public decimal Uren { get; set; }
 
         [Required(ErrorMessage = "Gelieve een bruto maandloon op te geven")]
         [Display(Name = "Bruto maandloon (fulltime)")]
         [Range(0, double.MaxValue, ErrorMessage = "Gelieve een positief getal voor het bruto maandloon op te geven.")]
-        public double BrutoMaandloonFulltime { get; set; }
+        public decimal BrutoMaandloonFulltime { get; set; }
 
-        public double Bedrag { get; set; }
+        public decimal Bedrag { get; set; }
         #endregion
 
         #region Constructors
