+++ src/KairosWeb_Groep6/Models/KairosViewModels/Baten/SubsidieViewModel.cs
@@ -7,6 +7,7 @@ namespace KairosWeb_Groep6.Models.KairosViewModels.Baten
 {
     public class SubsidieViewModel
     {
+        #region Properties
         [HiddenInput]
         public int Id { get; set; }
 
@@ -19,8 +20,10 @@ namespace KairosWeb_Groep6.Models.KairosViewModels.Baten
         [Display(Name = "Jaarbedrag")]
         [Required(ErrorMessage = "Gelieve het bedrag op te geven.")]
         [Range(0, double.MaxValue, ErrorMessage = "Gelieve een positief getal voor het bedrag op te geven.")]
-        public double Bedrag { get; set; }
+        public decimal Bedrag { get; set; }
+        #endregion
 
+        #region Constructors        
         public SubsidieViewModel()
         {
             Type = Type.Baat;
@@ -33,5 +36,6 @@ namespace KairosWeb_Groep6.Models.KairosViewModels.Baten
             Id = subsidie.Id;
             Bedrag = subsidie.Bedrag;
         }
+        #endregion
     }
 }
